[AF-2328] Moving BasicAuthorizationPrincipal.java to uberfire-security-backend
