+++ uberfire-backend/uberfire-backend-server/src/main/java/org/uberfire/backend/server/security/adapter/GroupAdapterAuthorizationSource.java
@@ -36,9 +36,9 @@ import org.jboss.errai.security.shared.api.Group;
 import org.jboss.errai.security.shared.api.GroupImpl;
 import org.jboss.errai.security.shared.api.Role;
 import org.jboss.errai.security.shared.api.RoleImpl;
-import org.uberfire.backend.server.security.BasicAuthorizationPrincipal;
 import org.uberfire.backend.server.security.RoleRegistry;
 import org.uberfire.security.authz.adapter.GroupsAdapter;
+import org.uberfire.security.backend.BasicAuthorizationPrincipal;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.unmodifiableList;
