+++ uberfire-backend/uberfire-backend-server/src/test/java/org/uberfire/backend/server/security/adapter/GroupAdapterAuthorizationSourceTest.java
@@ -30,7 +30,7 @@ import javax.management.remote.JMXPrincipal;
 import javax.security.auth.Subject;
 
 import org.junit.Test;
-import org.uberfire.backend.server.security.BasicAuthorizationPrincipal;
+import org.uberfire.security.backend.BasicAuthorizationPrincipal;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
